requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0
python-multipart==0.0.18
Pillow==10.2.0
